target/mips: Support direct execution from cache
The firmware for the XDuoo X3 wants to be able to execute directly
from cache while it writes to the underlying RAM via the
uncached-access address. Add some support for this.

This is pretty hacky and probably not upstreamable as-is.
Among other things, we don't restrict it to only the jz4760.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>