hw/mips: Create a skeleton model of the JZ4760 SoC
Create a skeleton model of the JZ4760 MIPS SoC, as
used in the XDuoo X3 music player.

This stubs out all the SoC's devices, and has a very minimal
bootrom, which relies on the board code loading the code to
be booted into the SDRAM for us. It is capable of starting to
execute the XDuoo-X3 rockbox code from NAND, though of course
it very quickly gets into trouble because of missing devices.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>