hw/mips/jz4760: Add UARTs
The JZ4760 UARTs are 16550-compatible; add them.
For the moment, we just stub out the extra 0x10 bytes of
registers each UART has at the SoC level. Unfortunately
QEMU's 16550 model has still not been QoMified and it is
not easy to extend it to support extra registers as part
of a UART model.

This is sufficient to make
 qemu-system-mipsel -M xduoo-x3  -bios rockbox.bin -serial null -serial stdio

print Rockbox's initial "SPL Stage 1" banner to the serial port.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>